+++ io/include/traccc/io/conditions_config.hpp
@@ -19,7 +19,7 @@ namespace traccc {
 
 /// Type describing the conditions configuration of a detector module
 struct conditions_data_config {
-    float threshold = 0.f;
+    double threshold = 0.f;
     vector2 shift{0.f, 0.f};
 };
